+++ pyuvsim/utils.py
@@ -57,18 +57,14 @@ class progsteps:
         self.step = step
         self.curval = -1
 
-    def update(self, count, mem=None):
+    def update(self, count):
         """
         Update the progress bar.
 
         Parameters
         ----------
-
         count : int
             Current value of counter
-        mem : float (optional)
-            Maximum current per-node memory usage in GB.
-            If provided, this information will also be printed.
         """
         if count >= self.curval + self.step:
             doprint = False
@@ -78,12 +74,9 @@ class progsteps:
             if doprint:
                 ostr = "{:0.2f}% completed. {:0.3f} minutes elapsed".format(
                     (count / self.maxval) * 100., (pytime.time() - self.t0) / 60.)
-                if mem is not None:
-                    ostr += ", MaxRSS = {:0.3f} GB\n".format(mem)
-                else:
-                    ostr += '\n'
+                ostr += '\n'
                 print(ostr)
-            sys.stdout.flush()
+                sys.stdout.flush()
 
     def finish(self):
         self.update(self.maxval)